refactor: move UDF code to dedicated module
